fixed y_missing bug

+++ tests/integration_tests/test_dashboards.py
@@ -99,7 +99,7 @@ def get_catboost_classifier():
                         labels=['Not survived', 'Survived'],
                         idxs=test_names)
 
-    X_cats, y_cats = explainer.X_merged, explainer.y
+    X_cats, y_cats = explainer.X_merged, explainer.y.astype("int")
     model = CatBoostClassifier(iterations=5, verbose=0).fit(X_cats, y_cats, cat_features=[5, 6, 7])
     explainer = ClassifierExplainer(model, X_cats, y_cats, idxs=X_test.index)
     explainer.calculate_properties(include_interactions=False)
